refactor(entries$): move 'sortEntriesByLastUpdated$' type to 'entries$' common module
